requirements.txt
tensorflow
requests
opencv-python
Pillow
fastapi
uvicorn
keras
opencv-python-headless
huggingface_hub